requirements.txt
google-cloud-vision
flask==0.10.1